[ios_notifications] Add checkSystemNotificationPermissionsThunk
