--- shared/src/androidMain/kotlin/ch/dreipol/multiplatform/reduxsample/shared/permissions/Notifications.kt
@@ -1,1 +1,0 @@
-package ch.dreipol.multiplatform.reduxsample.shared.permissions
