+++ shared/src/commonMain/kotlin/ch/dreipol/multiplatform/reduxsample/shared/redux/ReduxSampleApp.kt
@@ -4,6 +4,8 @@ import ch.dreipol.dreimultiplatform.reduxkotlin.presenterEnhancer
 import ch.dreipol.dreimultiplatform.uiDispatcher
 import ch.dreipol.multiplatform.reduxsample.shared.redux.middleware.*
 import ch.dreipol.multiplatform.reduxsample.shared.redux.reducer.rootReducer
+import ch.dreipol.multiplatform.reduxsample.shared.redux.thunk.initialNavigationThunk
+import ch.dreipol.multiplatform.reduxsample.shared.redux.thunk.loadSavedSettingsThunk
 import ch.dreipol.multiplatform.reduxsample.shared.utils.AppLanguage
 import ch.dreipol.multiplatform.reduxsample.shared.utils.fromLocale
 import org.reduxkotlin.applyMiddleware
