+++ shared/src/commonMain/kotlin/ch/dreipol/multiplatform/reduxsample/shared/redux/middleware/NotificationPermissionMiddleware.kt
@@ -1,13 +1,12 @@
 package ch.dreipol.multiplatform.reduxsample.shared.redux.middleware
 
 import ch.dreipol.multiplatform.reduxsample.shared.redux.AppState
-import ch.dreipol.multiplatform.reduxsample.shared.redux.OnboardingScreen
 import ch.dreipol.multiplatform.reduxsample.shared.redux.actions.NavigationAction
 import org.reduxkotlin.middleware
 
 fun notificationPermissionMiddleware() = middleware<AppState> { store, next, action ->
-    if (action == NavigationAction.ONBOARDING_NEXT && store.state.navigationState.currentScreen == OnboardingScreen(3)) {
-        print("Check permissions")
+    when (action) {
+        NavigationAction.ONBOARDING_NEXT -> print("check permission")
     }
     next(action)
 }
