+++ shared/src/commonMain/kotlin/ch/dreipol/multiplatform/reduxsample/shared/redux/thunk/Thunks.kt
@@ -1,11 +1,11 @@
-package ch.dreipol.multiplatform.reduxsample.shared.redux
+package ch.dreipol.multiplatform.reduxsample.shared.redux.thunk
 
 import ch.dreipol.dreimultiplatform.defaultDispatcher
 import ch.dreipol.dreimultiplatform.kermit
 import ch.dreipol.multiplatform.reduxsample.shared.database.*
-import ch.dreipol.multiplatform.reduxsample.shared.delight.NotificationSettings
 import ch.dreipol.multiplatform.reduxsample.shared.delight.Settings
 import ch.dreipol.multiplatform.reduxsample.shared.network.ServiceFactory
+import ch.dreipol.multiplatform.reduxsample.shared.redux.AppState
 import ch.dreipol.multiplatform.reduxsample.shared.redux.actions.*
 import ch.dreipol.multiplatform.reduxsample.shared.ui.DisposalCalendarEntry
 import ch.dreipol.multiplatform.reduxsample.shared.ui.DisposalCalendarMonth
@@ -18,7 +18,7 @@ import org.reduxkotlin.Thunk
 
 val networkAndDbScope = CoroutineScope(defaultDispatcher)
 
-private fun executeNetworkOrDbAction(action: suspend () -> Unit) {
+internal fun executeNetworkOrDbAction(action: suspend () -> Unit) {
     networkAndDbScope.launch {
         try {
             action.invoke()
@@ -105,54 +105,6 @@ fun loadSavedSettingsThunk(): Thunk<AppState> = { dispatch, _, _ ->
     }
 }
 
-fun setRemindTimeThunk(remindTime: RemindTime): Thunk<AppState> = { dispatch, getState, _ ->
-    val notification = getState.invoke().settingsState.notificationSettings?.firstOrNull()
-    notification?.let {
-        executeNetworkOrDbAction {
-            setNotificationSettings(it.copy(remindTime = remindTime))
-            dispatch(loadSavedSettingsThunk())
-        }
-    }
-    Unit
-}
-
-fun addOrRemoveNotificationThunk(): Thunk<AppState> = { dispatch, getState, _ ->
-    val settingsState = getState.invoke().settingsState
-    val notificationSettings = settingsState.notificationSettings
-    val notification = notificationSettings?.firstOrNull()
-    val defaultRemindTime = settingsState.settings?.defaultRemindTime ?: SettingsDataStore.defaultRemindTime
-    executeNetworkOrDbAction {
-        if (notification == null) {
-            setNotificationSettings(createNotification(SettingsDataStore.defaultShownDisposalTypes, defaultRemindTime))
-        } else {
-            setNotificationSettings(null)
-        }
-        dispatch(loadSavedSettingsThunk())
-    }
-}
-
-fun addOrRemoveNotificationThunk(disposalType: DisposalType): Thunk<AppState> = { dispatch, getState, _ ->
-    val settingsState = getState.invoke().settingsState
-    val notificationSettings = settingsState.notificationSettings
-    val settings = settingsState.settings
-    val notification = if (notificationSettings == null || notificationSettings.isEmpty()) {
-        createNotification(emptyList(), settings?.defaultRemindTime ?: SettingsDataStore.defaultRemindTime)
-    } else {
-        notificationSettings.first()
-    }
-    val disposalTypes = notification.disposalTypes.toMutableList()
-    if (disposalTypes.contains(disposalType)) {
-        disposalTypes.remove(disposalType)
-    } else {
-        disposalTypes.add(disposalType)
-    }
-    val updatedNotification = notification.copy(disposalTypes = disposalTypes.toList())
-    executeNetworkOrDbAction {
-        SettingsDataStore().insertOrUpdate(updatedNotification)
-        dispatch(loadSavedSettingsThunk())
-    }
-}
-
 fun setNewZipThunk(zip: Int): Thunk<AppState> = { dispatch, getState, _ ->
     val settingsState = getState.invoke().settingsState
     val showDisposalTypes = settingsState.settings?.showDisposalTypes ?: SettingsDataStore.defaultShownDisposalTypes
@@ -197,7 +149,7 @@ fun saveOnboardingThunk(): Thunk<AppState> = { dispatch, getState, _ ->
     }
     executeNetworkOrDbAction {
         saveSettings(settings)
-        setNotificationSettings(notification)
+        setNotificationSettings(notification, dispatch)
         dispatch(loadSavedSettingsThunk())
     }
 }
@@ -220,19 +172,9 @@ fun setNewAppLanguageThunk(appLanguage: AppLanguage, platformSpecificAction: () 
         }
     }
 
-private fun createNotification(disposalTypes: List<DisposalType>, remindTime: RemindTime): NotificationSettings {
-    return NotificationSettings(SettingsDataStore.UNDEFINED_ID, disposalTypes, remindTime)
-}
-
 private fun saveSettings(settings: Settings) {
     var settingsToSave = settings
     val settingsDataStore = SettingsDataStore()
     settingsDataStore.getSettings()?.let { settingsToSave = settings.copy(id = it.id) }
     settingsDataStore.insertOrUpdate(settingsToSave)
-}
-
-private fun setNotificationSettings(notificationSettings: NotificationSettings?) {
-    val settingsDataStore = SettingsDataStore()
-    settingsDataStore.deleteNotificationSettings()
-    notificationSettings?.let { settingsDataStore.insertOrUpdate(it) }
 }
