+++ shared/src/commonMain/kotlin/ch/dreipol/multiplatform/reduxsample/shared/ui/ZipSettingsView.kt
@@ -1,7 +1,7 @@
 package ch.dreipol.multiplatform.reduxsample.shared.ui
 
 import ch.dreipol.dreimultiplatform.reduxkotlin.rootDispatch
-import ch.dreipol.multiplatform.reduxsample.shared.redux.setNewZipThunk
+import ch.dreipol.multiplatform.reduxsample.shared.redux.thunk.setNewZipThunk
 
 data class ZipSettingsViewState(
     val headerViewState: HeaderViewState = HeaderViewState("settings_zip"),
